Removing clusterManagerAdapter from topologyDescriptor constructor Removing CorfuReplicationClusterManagerBaseAdapter
